Fixed Gjs warning
Tested on arch linux (gnome-shell 49.0.2).
Using the recommended js method to parse Uint8Array as described by the
warning below.
Gjs-WARNING **: 02:10:33.069: Some code called array.toString() on a Uint8Array instance. Previously this would have interpreted the bytes of the array as a string, but that is nonstandard. In the future this will return the bytes as comma-separated digits. For the time being, the old behavior has been preserved, but please fix your code anyway to use TextDecoder.